Inspect middleware init parameters before passing in lifespan argument
